Formula pakage

+++ app/FormulaPakage/router/selected_files.py
@@ -7,7 +7,7 @@ from app.TablePakage.model.database import get_db
 from app.TablePakage.model.parameter_schema import ParameterSchema
 
 from ..model.selected_file import SelectedFile
-from ..schema.selected_file_schema import SelectedFileSchemaCreate, SelectedFileSchemaUpdate, SelectedFileSchemaResponse
+from ..schema.selected_file_schema import SelectedFileSchemaGet, SelectedFileSchemaCreate, SelectedFileSchemaUpdate, SelectedFileSchemaResponse
 
 from .fields_of_view import FIELDS_OF_VIEW_PATTERN
 
@@ -19,6 +19,7 @@ import uuid
 router = APIRouter(prefix="/selected_file", tags=["SelectedFiles"])
 
 STORAGE_PATH = ("./static/param_files")
+os.makedirs(STORAGE_PATH, exist_ok=True)
 
 def generate_unique_filename(original_filename: str) -> str:
     ext = Path(original_filename).suffix
@@ -26,7 +27,7 @@ def generate_unique_filename(original_filename: str) -> str:
     return unique_name
 
 
-@router.put("/update/{node_id}", response_model=SelectedFileSchemaResponse, description="Занесение данных в таблицу")
+@router.put("/update/{node_id}", description="Занесение данных в таблицу") # response_model=SelectedFileSchemaResponse, 
 async def update(
     node_id: int,
     schema_update: SelectedFileSchemaUpdate,
@@ -34,19 +35,20 @@ async def update(
 ):
     try:
         result = await db.execute(select(SelectedFile).where(SelectedFile.id == node_id))
-        existing_node = result.one_or_none()
+        existing_node = result.scalar_one_or_none()
         if not existing_node:
             raise HTTPException(status_code=404, detail=f"Отсутствует запись в SelectedFile с id: {node_id}")
         
         for key, value in schema_update.dict(exclude_unset=True).items():
             setattr(existing_node, key, value)
-        await db.refresh(existing_node)
+        
         await db.commit()
+        await db.refresh(existing_node)
 
         # Сборка шаблона
         stmt = select(ParameterSchema).where(ParameterSchema.id == existing_node.parametr_schema_id)
         res = await db.execute(stmt)
-        param = res.one_or_none()
+        param = res.scalar_one_or_none()
         if not param:
             raise HTTPException(status_code=404, detail=f"Отсутствует параметр с id: {existing_node.parametr_schema_id}")
         selected_file_result = {'fields': []}
@@ -55,13 +57,12 @@ async def update(
             'file_path': existing_node.file_path,
             'file_url': existing_node.file_url,
             'name': existing_node.name,
-            'description': existing_node.description,
             'parametr_schema_name': param.name,
             'parametr_schema_id': param.id
         }
         for field in FIELDS_OF_VIEW_PATTERN['selected_file']['fields']:
             if field['field'] in data and data[field['field']] is not None:
-                field['value'] = selected_file[field['field']]
+                field['value'] = data[field['field']]
                 selected_file_result['fields'].append(field)
                 continue
                 
@@ -75,7 +76,7 @@ async def update(
         await db.rollback()
         raise HTTPException(status_code=500, detail=f"Ошибка при обновлении записи с id: {node_id} в таблице SelectedFile: {str(e)}")
 
-@router.put("/upload/{node_id}", response_model=SelectedFileSchemaResponse, description="Загрузка файла к параметру")
+@router.put("/upload/{node_id}", description="Загрузка файла к параметру")
 async def upload(
     node_id: int,
     sel_file: UploadFile,
@@ -83,7 +84,7 @@ async def upload(
 ):
     try:
         result = await db.execute(select(SelectedFile).where(SelectedFile.id == node_id))
-        existing_node = result.one_or_none()
+        existing_node = result.scalar_one_or_none()
         if not existing_node:
             raise HTTPException(status_code=404, detail=f"Отсутствует запись в SelectedFile с id: {node_id}")
         
@@ -103,13 +104,14 @@ async def upload(
         existing_node.file_url = file_url
         existing_node.content_type = file_fromat
 
-        await db.refresh(existing_node)
-        await db.commit()
         
+        await db.commit()
+        await db.refresh(existing_node)
+
         # Сборка шаблона
         stmt = select(ParameterSchema).where(ParameterSchema.id == existing_node.parametr_schema_id)
         res = await db.execute(stmt)
-        param = res.one_or_none()
+        param = res.scalar_one_or_none()
         if not param:
             raise HTTPException(status_code=404, detail=f"Отсутствует параметр с id: {existing_node.parametr_schema_id}")
         selected_file_result = {'fields': []}
@@ -118,13 +120,13 @@ async def upload(
             'file_path': existing_node.file_path,
             'file_url': existing_node.file_url,
             'name': existing_node.name,
-            'description': existing_node.description,
+            'content_type': existing_node.content_type,
             'parametr_schema_name': param.name,
             'parametr_schema_id': param.id
         }
         for field in FIELDS_OF_VIEW_PATTERN['selected_file']['fields']:
             if field['field'] in data and data[field['field']] is not None:
-                field['value'] = selected_file[field['field']]
+                field['value'] = data[field['field']]
                 selected_file_result['fields'].append(field)
                 continue
                 
@@ -135,16 +137,16 @@ async def upload(
     except HTTPException:
         raise
     except Exception as e:
-        db.rollback()
+        await db.rollback()
         raise HTTPException(status_code=500, detail=f"Ошибка при загрузке файла в запись с id: {node_id} в таблицу SelectedFile: {str(e)}")
 
 
 
-@router.post("/add_param/{param_id}", response_model=SelectedFileSchemaResponse, description="Создание записи для загрузки файла")
+@router.post("/add_param/{param_id}", description="Создание записи для загрузки файла")
 async def add_param_to_selected_file(param_id: int, db: AsyncSession = Depends(get_db)):
     try:
         result = await db.execute(select(ParameterSchema).where(ParameterSchema.id == param_id))
-        param = result.one_or_none()
+        param = result.scalar_one_or_none()
         if not param:
             raise HTTPException(status_code=404, detail=f"Отсутствует параметр с id: {param_id}")
         new_node = SelectedFile(parametr_schema_id=param_id)
@@ -158,14 +160,14 @@ async def add_param_to_selected_file(param_id: int, db: AsyncSession = Depends(g
             'id': new_node.id,
             'file_path': new_node.file_path,
             'file_url': new_node.file_url,
+            'content_type': new_node.content_type,
             'name': new_node.name,
-            'description': new_node.description,
             'parametr_schema_name': param.name,
             'parametr_schema_id': param.id
         }
         for field in FIELDS_OF_VIEW_PATTERN['selected_file']['fields']:
             if field['field'] in data and data[field['field']] is not None:
-                field['value'] = selected_file[field['field']]
+                field['value'] = data[field['field']]
                 selected_file_result['fields'].append(field)
                 continue
                 
@@ -178,30 +180,33 @@ async def add_param_to_selected_file(param_id: int, db: AsyncSession = Depends(g
         await db.rollback()
         raise HTTPException(status_code=500, detail=f"Ошибка при добавлении параметра с id: {param_id} в таблицу SelectedFile: {str(e)}")
 
-@router.delete("/delete_file/{node_id}", response_model=SelectedFileSchemaResponse, description="Удаление файла")  
+@router.delete("/delete_file/{node_id}", description="Удаление файла")  
 async def delete_file(
     node_id: int,
     db: AsyncSession = Depends(get_db)
 ):
     try:
         result = await db.execute(select(SelectedFile).where(SelectedFile.id == node_id))
-        existing_node = result.one_or_none()
+        existing_node = result.scalar_one_or_none()
         if not existing_node:
             raise HTTPException(status_code=404, detail=f"Отсутствует запись в SelectedFile с id: {node_id}")
 
-        exist_file = os.path.exists(existing_node.file_path)
-        if not exist_file:
-            raise HTTPException(status_code=404, detail=f"Отсутствует файл в SelectedFile с id: {node_id}")
+        if existing_node.file_path:
+            exist_file = os.path.exists(existing_node.file_path)
+            if not exist_file:
+                raise HTTPException(status_code=404, detail=f"Отсутствует файл в SelectedFile с id: {node_id}")
         
-        os.remove(file_path)
+            os.remove(existing_node.file_path)
         existing_node.file_path = None
-        await db.refresh(existing_node)
+        existing_node.file_url = None
+        
         await db.commit()
+        await db.refresh(existing_node)
 
         # Сборка шаблона
         stmt = select(ParameterSchema).where(ParameterSchema.id == existing_node.parametr_schema_id)
         res = await db.execute(stmt)
-        param = res.one_or_none()
+        param = res.scalar_one_or_none()
         if not param:
             raise HTTPException(status_code=404, detail=f"Отсутствует параметр с id: {existing_node.parametr_schema_id}")
         selected_file_result = {'fields': []}
@@ -210,13 +215,12 @@ async def delete_file(
             'file_path': existing_node.file_path,
             'file_url': existing_node.file_url,
             'name': existing_node.name,
-            'description': existing_node.description,
             'parametr_schema_name': param.name,
             'parametr_schema_id': param.id
         }
         for field in FIELDS_OF_VIEW_PATTERN['selected_file']['fields']:
             if field['field'] in data and data[field['field']] is not None:
-                field['value'] = selected_file[field['field']]
+                field['value'] = data[field['field']]
                 selected_file_result['fields'].append(field)
                 continue
                 
@@ -229,25 +233,25 @@ async def delete_file(
         raise
     except Exception as e:
         await db.rollback()
-        raise HTTPException(status_code=500, detail=f"Ошибка при удалении файла в SelectedFile с id: {node_id}")
+        raise HTTPException(status_code=500, detail=f"Ошибка при удалении файла в SelectedFile с id: {node_id} - {e}")
 
-@router.delete("/delete_node/{node_id}", response_model=SelectedFileSchemaResponse, description="Удаление записи с SelectedFile")  
+@router.delete("/delete_node/{node_id}", description="Удаление записи с SelectedFile")  
 async def delete_node(
     node_id: int,
     db: AsyncSession = Depends(get_db)
 ):
     try:
         result = await db.execute(select(SelectedFile).where(SelectedFile.id == node_id))
-        existing_node = result.one_or_none()
+        existing_node = result.scalar_one_or_none()
         if not existing_node:
             raise HTTPException(status_code=404, detail=f"Отсутствует запись в SelectedFile с id: {node_id}")
-
-        exist_file = os.path.exists(existing_node.file_path)
-        if not exist_file:
-            raise HTTPException(status_code=404, detail=f"Отсутствует файл в SelectedFile с id: {node_id}")
+        if existing_node.file_path:
+            exist_file = os.path.exists(existing_node.file_path)
+            if not exist_file:
+                raise HTTPException(status_code=404, detail=f"Отсутствует файл в SelectedFile с id: {node_id}")
         
-        os.remove(file_path)
-        existing_node.file_path = None
+            os.remove(file_path)
+            existing_node.file_path = None
         await db.delete(existing_node)
         await db.commit()
 
@@ -259,7 +263,7 @@ async def delete_node(
         raise HTTPException(status_code=500, detail=f"Ошибка при удалении записи в SelectedFile с id: {node_id}: {e}")
 
 
-@router.get("/get_selected_file/{node_id}", response_model=SelectedFileSchemaResponse, description="Получение данных о загруженном файле к параметру")
+@router.get("/get_selected_file/{node_id}", description="Получение данных о загруженном файле к параметру")
 async def get_selected_file(node_id: int, db: AsyncSession = Depends(get_db)):
     try:
         stmt = select(
@@ -277,13 +281,13 @@ async def get_selected_file(node_id: int, db: AsyncSession = Depends(get_db)):
             raise HTTPException(status_code=404, detail=f"Отсутствует запись в SelectedFile с id: {node_id}")
         selected_file_result = {'fields': []}
         data = {
-            'id': row.id,
-            'file_path': row.file_path,
-            'file_url': row.file_url,
-            'name': row.name,
-            'content_type': row.content_type,
-            'parametr_schema_name': row.parametr_schema_name,
-            'parametr_schema_id': row.parametr_schema_id
+            'id': selected_file.id,
+            'file_path': selected_file.file_path,
+            'file_url': selected_file.file_url,
+            'name': selected_file.name,
+            'content_type': selected_file.content_type,
+            'parametr_schema_name': selected_file.parametr_schema_name,
+            'parametr_schema_id': selected_file.parametr_schema_id
         }
         for field in FIELDS_OF_VIEW_PATTERN['selected_file']['fields']:
             if field['field'] in data and data[field['field']] is not None:
@@ -299,7 +303,7 @@ async def get_selected_file(node_id: int, db: AsyncSession = Depends(get_db)):
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Ошибка при получении записи в SelectedFile с id: {node_id}: {e}")
 
-@router.get("/get_selected_files", response_model=List[SelectedFileSchemaResponse], description="Получение данных о всех загруженных файлах")
+@router.get("/get_selected_files", response_model=List[SelectedFileSchemaGet], description="Получение данных о всех загруженных файлах")
 async def get_selected_files(db: AsyncSession = Depends(get_db)):
     try:
         res = []
@@ -312,17 +316,17 @@ async def get_selected_files(db: AsyncSession = Depends(get_db)):
             ParameterSchema.id.label('parametr_schema_id')
         ).join(ParameterSchema).where(ParameterSchema.id == SelectedFile.parametr_schema_id)
         result = await db.execute(stmt)
-        selected_files = result.scalars().all()
+        selected_files = result.fetchall()
         if not selected_files:
             return res
         for files in selected_files:
             data = {
-                'id': row.id,
-                'file_path': row.file_path,
-                'file_url': row.file_url,
-                'name': row.name,
-                'parametr_schema_name': row.parametr_schema_name,
-                'parametr_schema_id': row.parametr_schema_id
+                'id': files.id,
+                'file_path': files.file_path,
+                'file_url': files.file_url,
+                'name': files.name,
+                'parametr_schema_name': files.parametr_schema_name,
+                'parametr_schema_id': files.parametr_schema_id
             }
             res.append(data)
         return res
